refactor: clean code
