+++ react-components/src/components/Card/Card.module.scss
@@ -10,8 +10,8 @@
 
 .imageProduct {
   border-radius: 20px 20px 0 0;
-  width: 300px;
-  height: 310px;
+  height: 240px;
+  width: 210px;
 }
 
 .wrapperInfo {
