+++ react-components/src/components/CardsList/CardsList.module.scss
@@ -6,7 +6,7 @@
 
 .cards {
   display: grid;
-  grid-template-columns: repeat(auto-fill, minmax(330px, 1fr));
+  grid-template-columns: repeat(auto-fill, minmax(250px, 1fr));
   gap: 40px;
   margin: 0 auto;
 }
