+++ react-components/src/components/SearchBar/index.ts
