+++ react-components/src/scss/styles.scss
@@ -1,3 +1,5 @@
+@import url('https://fonts.googleapis.com/css2?family=EB+Garamond:wght@400;500;600;700&family=Libre+Baskerville:wght@400;700&family=Noto+Sans+Georgian:wght@100;200;300;400;500;600;700;800;900&family=Roboto:ital,wght@0,100;0,300;0,400;0,500;0,700;0,900;1,100;1,300;1,400;1,500;1,700;1,900&family=Rubik+Bubbles&display=swap');
+
 * {
   box-sizing: border-box;
   font-family: var(--font-family-roboto);
@@ -8,6 +10,7 @@
 
 body {
   background-color: #eae9ee;
+  font-family: 'EB Garamond', serif;
 }
 
 a {
